+++ application/controllers/Welcome.php
@@ -29,5 +29,32 @@ class Welcome extends Application
 
 		$this->render();
 	}
+	
+	/**
+	* Random quote generation
+	*/
+	public function random()
+    {
+        
+        // this is the view we want shown
+        $this->data['pagebody'] = 'homepage';
+
+        // build the list of authors, to pass on to our view
+        $source = $this->quotes->all();        
+        
+        $authors = array ();
+		
+		// randomly select a person and quote
+        $record = $source[mt_rand(0, count($source) - 1)];
+		
+		// display the person and quote
+		$authors[] = array('who' => $record['who'], 'mug' => $record['mug'], 'href' => $record['where'], 'what' => $record['what']);
+
+        $this->data['authors'] = $authors;
+
+        $this->render();
+    }
+	
+	
 
 }